License, readme, warnings, cleanup
