+++ lib/compressTextures.js
@@ -7,6 +7,7 @@ var os = require('os');
 var path = require('path');
 var Promise = require('bluebird');
 var uuid = require('uuid');
+var getTempDirectory = require('./getTempDirectory');
 
 Jimp.prototype.getBufferAsync = Promise.promisify(Jimp.prototype.getBuffer);
 var fsExtraEnsureDir = Promise.promisify(fsExtra.ensureDir);
@@ -145,7 +146,16 @@ function compressTextures(gltf, options) {
             if (defined(jimpImage)) {
                 width = jimpImage.bitmap.width;
                 height = jimpImage.bitmap.height;
-                resize = resizeToPowerOfTwo && (!CesiumMath.isPowerOfTwo(width) || !CesiumMath.isPowerOfTwo(height));
+                var isPowerOfTwo = CesiumMath.isPowerOfTwo(width) && CesiumMath.isPowerOfTwo(height);
+                var isSquare = (width === height);
+                resize = resizeToPowerOfTwo && !isPowerOfTwo;
+
+                // Log warnings if the image will be resized. This applies to all compression formats.
+                if ((format === 'pvrtc1' || format === 'pvrtc2') && (!isSquare || !isPowerOfTwo)) {
+                    console.log('Warning: ' + imageId + ' will be resized to the lower square power-of-two for pvrtc texture compression. To prevent this warning in the future resize any images beforehand.');
+                } else if (!isPowerOfTwo) {
+                    console.log('Warning: ' + imageId + ' will be resized to the lower power-of-two for texture compression. To prevent this warning in the future resize any images beforehand.');
+                }
             }
 
             var promise;
@@ -218,11 +228,6 @@ function replaceImage(image) {
     };
 }
 
-function getTempDirectory() {
-    var tempDirectory = os.tmpdir();
-    var randomId = uuid.v4();
-    return path.join(tempDirectory, randomId);
-}
 
 function getTempImagePath(tempDirectory, extension) {
     var randomId = uuid.v4();
