+++ lib/parseArguments.js
@@ -137,12 +137,6 @@ function parseArguments(args) {
                 nargs: 1,
                 type: 'string'
             },
-            'texcomp.enable': {
-                default : false,
-                describe: 'Compress textures.',
-                group: 'Options: Texture Compression',
-                type: 'boolean'
-            },
             'texcomp.format': {
                 choices: ['pvrtc1', 'pvrtc2', 'etc1', 'etc2', 'astc', 'dxt1', 'dxt3', 'dxt5', 'crunch-dxt1', 'crunch-dxt3', 'crunch-dxt5'],
                 describe: 'The compressed texture format.',
